Update sandbox stories
