+++ code/frameworks/nextjs/template/stories_nextjs-default-ts/Router.stories.tsx
@@ -19,10 +19,12 @@ function Component() {
       name: 'Prefetch',
     },
     {
+      // @ts-expect-error (old API)
       cb: () => router.push('/push-html', { forceOptimisticNavigation: true }),
       name: 'Push HTML',
     },
     {
+      // @ts-expect-error (old API)
       cb: () => router.replace('/replaced-html', { forceOptimisticNavigation: true }),
       name: 'Replace',
     },
