+++ test-storybooks/portable-stories-kitchen-sink/nextjs/package.json
@@ -14,7 +14,7 @@
     "playwright": "echo 'No playwright setup yet'"
   },
   "dependencies": {
-    "next": "14.1.0",
+    "next": "^14.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
@@ -36,7 +36,6 @@
     "eslint-plugin-storybook": "^0.6.15",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "next": "14.1.0",
     "storybook": "^8.0.0",
     "typescript": "^5.2.2"
   },
